requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# API
fastapi

# utilities
six>=1.14
google-cloud-translate==2.0.1
